feat(bevy): 英雄自动喝药（#196，SetAutoPotValue/Item 协议修正 + 阈值按钮）
- SharedRust：修正 client::hero::SetAutoPotValue（value u8→u32，C# uint，5 字节）
  与 client::hero::SetAutoPotItem（grid u32→u8、item_index u64→i32，C# 5 字节）；
  修正 server::miscellaneous::SetAutoPotItem（slot/item_id i32+i32 → grid u8+item_index i32）
- Client-Macroquad：按新字段同步（value u32 / item_index i32）
- Client-Bevy：解码 S.SetAutoPotValue/SetAutoPotItem → ServerEvent；英雄对话框新增
  自动药 HP/MP 阈值循环按钮（0/30/50/70/90，Title 560/563）→ C.SetAutoPotValue{stat 12/13}
- 新增 SharedRust SetAutoPotItem round-trip 测试（163 单测全过）

验证：SharedRust 163 单测；Client-Bevy cargo check + 61+6 测试 + mock 实机无 panic。

+++ Client-Macroquad/src/network/client.rs
@@ -1433,7 +1433,7 @@ fn handle_outbound_event<S: Write>(stream: &mut S, event: NetworkEvent) -> Resul
         NetworkEvent::SetHeroAutoPotValue { pot_type, value } => {
             let packet = client::hero::SetAutoPotValue {
                 stat: pot_type,
-                value: (value as u8).min(100),
+                value: (value as u32).min(100),
             };
             serialize_packet(stream, &packet)?;
             tracing::debug!("📤 SetAutoPotValue: stat={}, value={}", pot_type, value);
@@ -1442,7 +1442,7 @@ fn handle_outbound_event<S: Write>(stream: &mut S, event: NetworkEvent) -> Resul
         NetworkEvent::SetHeroAutoPotItem { item_id } => {
             let packet = client::hero::SetAutoPotItem {
                 grid: 0,
-                item_index: item_id as u64,
+                item_index: item_id as i32,
             };
             serialize_packet(stream, &packet)?;
             tracing::debug!("📤 SetAutoPotItem: item_id={}", item_id);
